backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
ultralytics>=8.1
torch>=2.0
opencv-python-headless>=4.8
numpy>=1.24
requests>=2.31
pydantic>=2.0
python-dotenv>=1.0
python-multipart>=0.0.6
